+++ attributes/default.rb
@@ -15,7 +15,10 @@ default['golang']['install_dir'] = '/usr/local'
 default['golang']['gopath'] = '/opt/go'
 default['golang']['gobin'] = '/opt/go/bin'
 default['golang']['scm'] = true
-default['golang']['scm_packages'] = %w(git mercurial bzr)
+default['golang']['scm_packages'] = value_for_platform(
+  'centos' => { '>= 8' => %w(git mercurial) },
+  'default' => %w(git mercurial bzr)
+)
 default['golang']['packages'] = []
 default['golang']['owner'] = 'root'
 default['golang']['group'] = 'root'